web/requirements.txt
Django==4.1
gunicorn==20.1.0
django-dotenv==1.4.2
psycopg2-binary==2.9.3
